moved ui to better match file logic
