+++ code/app/src/main/res/layout/fragment_questions_list.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".QuestionsFragment"
+    tools:context=".QuestionUI.QuestionsFragment"
     tools:listitem="@layout/fragment_questions" />
